# app/routes/_pages.$session_id.report/modules/loader.ts
import type { LoaderFunctionArgs } from "react-router";
import { api } from "~/libs/api";
import { notfound } from "~/libs/response";

export const loader = async ({ params, request }: LoaderFunctionArgs) => {
  if (!params.session_id) {
    return notfound();
  }

  const { data: report } = await api.GET(
    "/talksessions/{talkSessionID}/report",
    {
      headers: request.headers,
      params: {
        path: {
          talkSessionID: params.session_id,
        },
      },
    },
  );

  const { data } = await api.GET("/talksessions/{talkSessionID}/opinions", {
    headers: request.headers,
    params: {
      path: {
        talkSessionID: params.session_id,
      },
    },
  });

  const { data: position } = await api.GET(
    "/talksessions/{talkSessionID}/analysis",
    {
      headers: request.headers,
      params: {
        path: {
          talkSessionID: params.session_id,
        },
      },
    },
  );

  if (!data?.opinions) {
    return notfound();
  }

  return {
    position,
    opinions: data?.opinions,
    report: report?.report,
  };
};